+++ tests/testthat/test-chk_pattern.R
@@ -1,3 +1,3 @@
 test_that("scientific notation not applied when pattern checking numeric vars", {
-  expect_true(chk_pattern(10000000, "1[0]{7}"))
+  expect_true(chk_regex(10000000, "1[0]{7}"))
 })